Update product version string in PashHi and PasHiGUI
Product version string is the same in both prograns and has been altered
to just contain the actual product version, with no prefixed "Release"
text.

This was done to play nicely with the planned Deploy script that will
need to extract this version number to use in setup files and setup and
zip file names.