Make --verbosity command more flexible
There are now three basic verbosity states: `info`, `warning` and
`error`. Any combination of them, or none of them are supported. Set
notation is used to specify the required states, e.g `{info,warning}` or
`{}`.

The previous `normal`, `no-warn` and `quiet` states are now aliases for
relevant sets of verbosity states to provide backward compatibility. A
new `silent` alias was also added. `normal` and `no-warn` are made
deprecated, but `no-warn` is never reported as deprecated since it
switches off warnings!!

The TConfig.Verbosity property was changed from the former TVerbosity
enumerated type to the TVerbosityStates set type.

TParams was updated to read the revised states.

TConsole was overhauled completely to work with the new way verbosity
levels work.

The way TMain displays text was modified to take account of changes in
TConfig and TConsole.