Implement --config-show command line option
This option displays the commands that are specified in any config file.

Added support for parsing the option to TParams. Also added the option
to the black list of files that are not allowed in the config file.

Added a related property to TConfig.

Added code to TParams and TConfig to expose the value read from the
config file via a property of TConfig.

Updated TMain to write the config file commands to the console.